adding edit cabin form

+++ src/services/apiCabins.js
@@ -10,29 +10,46 @@ export async function getCabins() {
   return data;
 }
 
-/**https://gdvaljjgtxamgelngzoa.supabase.co/storage/v1/object/public/cabins-images/cabin-001.jpg */
-
-export async function createCabin(newCabin) {
+export async function createCabin(newCabin, id) {
+  console.log(id);
+  const hasImage = typeof newCabin?.image === "string";
+  console.log(hasImage);
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabins-images/${imageName}`;
-  const { data, error } = await supabase
-    .from("Cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+
+  const imagePath = hasImage
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabins-images/${imageName}`;
+
+  let query = supabase.from("Cabins");
+
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  } else {
+    query = supabase
+      .from("Cabins")
+      .update({ ...newCabin, image: imagePath })
+      .eq("id", id);
+  }
+
+  const { data, error } = await query.select().single();
+
   if (error) {
     throw new Error("Cabin can not be created");
   }
-  const { error: StorageError } = await supabase.storage
-    .from("cabins-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImage) {
+    const { error: StorageError } = await supabase.storage
+      .from("cabins-images")
+      .upload(imageName, newCabin.image);
 
-  if (StorageError) {
-    await supabase.from("Cabins").delete().eq("id", newCabin.id);
-    throw new Error("Storage Error,Cabin can not be created");
+    if (StorageError) {
+      await supabase.from("Cabins").delete().eq("id", newCabin.id);
+      throw new Error("Storage Error,Cabin can not be created");
+    }
   }
+
   return data;
 }
